+++ src/lib/utils/TypedObject.ts
@@ -1,9 +1,9 @@
-import { TupleOrArray } from '../../types/utils'
+import { ExtractObjectKeys, ExtractObjectValues, ExtractObjectEntries } from '../../types/utils'
 
 /**
  * Utility class for working with objects.
  *
- * Provides **static** methods for retrieving keys and values of an object.
+ * Provides **static** methods for retrieving keys and values of an object with addition of types.
  *
  * This class enhances type safety by providing better typings for object keys and values.
  */
@@ -13,27 +13,49 @@ export class TypedObject {
      * Returns the names of the enumerable string properties and methods of an object.
      *
      * Type parameters:
-     * - `K` (`TupleOrArray<string>`, defaults to `string[]`) -
-     * The type of the array containing the names of the enumerable string properties and methods.
      *
-     * @param {any} obj Object that contains the properties and methods.
-     * @returns {K} Array of names of the enumerable string properties and methods of the specified object.
+     * - `TObject` (`Record<string, any>`) - The object to get the object keys types from.
+     *
+     * @param {TObject} obj Object to get the keys from.
+     *
+     * @returns {Array<ExtractObjectKeys<TObject>>}
+     * Array of names of the enumerable string properties and methods of the specified object.
      */
-    public static keys<K extends TupleOrArray<string> = string[]>(obj: any): K {
-        return Object.keys(obj || {}) as K
+    public static keys<TObject extends Record<string, any>>(obj: TObject): ExtractObjectKeys<TObject>[] {
+        return Object.keys(obj || {})
     }
 
     /**
      * Returns an array of values of the enumerable properties of an object.
      *
      * Type parameters:
-     * - `V` (`TupleOrArray<any>`, defaults to `any[]`) -
-     * The type of the array containing the values of the enumerable properties.
      *
-     * @param {any} obj Object that contains the properties and methods.
-     * @returns {V} Array of values of the enumerable properties of the specified object.
+     * - `TObject` (`Record<string, any>`) - The object to get the object values types from.
+     *
+     * @param {TObject} obj Object to get the values from.
+     *
+     * @returns {Array<ExtractObjectValues<TObject>>}
+     * Array of values of the enumerable properties of the specified object.
      */
-    public static values<V extends TupleOrArray<any> = any[]>(obj: any): V {
-        return Object.values(obj || {}) as V
+    public static values<TObject extends Record<string, any>>(obj: TObject): ExtractObjectValues<TObject>[] {
+        return Object.values(obj || {})
+    }
+
+    /**
+     * Returns an array of entries (key-value pairs) of the enumerable properties of an object.
+     *
+     * Type parameters:
+     *
+     * - `TObject` (`Record<string, any>`) - The object to get the object entries types (key-value pairs types) from.
+     *
+     * @param {TObject} obj Object to get the entries from.
+     *
+     * @returns {ExtractObjectEntries<TObject>[]}
+     * Array of entries (key-value pairs) of the enumerable properties of the specified object.
+     */
+    public static entries<
+        TObject extends Record<string, any>
+    >(obj: TObject): ExtractObjectEntries<TObject>[] {
+        return Object.entries(obj || {})
     }
 }